LayoutTests/inspector/protocol/runtime-agent.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/protocol-test.js"></script>
<script>

function TestObject()
{
}

testObject = new TestObject();
testProperty = true;

function test()
{
    function callback(error, result)
    {
        var invalidObjectId = {id: -1, injectedScriptId: result.objectId.injectedScriptId};
        var testSuite = [
            ["RuntimeAgent", "evaluate", 'testObject', 'test', false],
            ["RuntimeAgent", "evaluate", 'testObject', 'test'],
            ["RuntimeAgent", "callFunctionOn", result.objectId, 'function() { this.assignedByCallFunctionOn = "callFunctionOn function works fine"; return this.assignedByCallFunctionOn; }'],
            ["RuntimeAgent", "getProperties", result.objectId, false],
            ["RuntimeAgent", "releaseObject", result.objectId],
            ["RuntimeAgent", "releaseObjectGroup", 'test']];

        InspectorTest.runProtocolTestSuite("RuntimeAgent", testSuite, ['id', 'requestId', 'objectId', 'className']);
    }

    RuntimeAgent.evaluate('window.testObject', 'test', false, callback);
}

</script>
</head>

<body onload="runTest()">
<p>
Protocol stability test. It is dumping request/response pairs of RuntimeAgent functions.
</p>

</body>
</html>
