LayoutTests/inspector/protocol/runtime-agent-expected.txt

Protocol stability test. It is dumping request/response pairs of RuntimeAgent functions.

-----------------------------------------------------------
RuntimeAgent.evaluate("testObject","test",false)

request:
{
    method : "Runtime.evaluate"
    params : {
        expression : "testObject"
        objectGroup : "test"
        includeCommandLineAPI : false
    }
    id : <number>
}

response:
{
    result : {
        result : {
            type : "object"
            objectId : <string>
            className : <string>
            description : "TestObject"
        }
    }
    id : <number>
}

-----------------------------------------------------------
RuntimeAgent.evaluate("testObject","test")

request:
{
    method : "Runtime.evaluate"
    params : {
        expression : "testObject"
        objectGroup : "test"
    }
    id : <number>
}

response:
{
    result : {
        result : {
            type : "object"
            objectId : <string>
            className : <string>
            description : "TestObject"
        }
    }
    id : <number>
}

-----------------------------------------------------------
RuntimeAgent.callFunctionOn(<string>,"function() { this.assignedByCallFunctionOn = \"callFunctionOn function works fine\"; return this.assignedByCallFunctionOn; }")

request:
{
    method : "Runtime.callFunctionOn"
    params : {
        objectId : <string>
        functionDeclaration : "function() { this.assignedByCallFunctionOn = "callFunctionOn function works fine"; return this.assignedByCallFunctionOn; }"
    }
    id : <number>
}

response:
{
    result : {
        result : {
            type : "string"
            value : "callFunctionOn function works fine"
        }
    }
    id : <number>
}

-----------------------------------------------------------
RuntimeAgent.getProperties(<string>,false)

request:
{
    method : "Runtime.getProperties"
    params : {
        objectId : <string>
        ignoreHasOwnProperty : false
    }
    id : <number>
}

response:
{
    result : {
        result : [
            {
                name : "assignedByCallFunctionOn"
                value : {
                    type : "string"
                    value : "callFunctionOn function works fine"
                }
            }
            {
                name : "__proto__"
                value : {
                    type : "object"
                    objectId : <string>
                    className : <string>
                    description : "TestObject"
                }
            }
        ]
    }
    id : <number>
}

-----------------------------------------------------------
RuntimeAgent.releaseObject(<string>)

request:
{
    method : "Runtime.releaseObject"
    params : {
        objectId : <string>
    }
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

-----------------------------------------------------------
RuntimeAgent.releaseObjectGroup("test")

request:
{
    method : "Runtime.releaseObjectGroup"
    params : {
        objectGroup : "test"
    }
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

===========================================================
Coverage for RuntimeAgent
{
    evaluate : "checked"
    callFunctionOn : "checked"
    getProperties : "checked"
    releaseObject : "checked"
    releaseObjectGroup : "checked"
}
